requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
black==20.8b1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
pdfminer.six==20200517
pdfplumber==0.5.23
Pillow==7.2.0
pycryptodome==3.9.8
PySocks==1.7.1
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
toml==0.10.1
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Wand==0.6.3
